requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.15
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
line-bot-sdk==3.14.2
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.0
pydantic==2.10.0
pydantic_core==2.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
yarl==1.17.2
requests-cache
